requirements.txt
flask
opencv-python
opencv-contrib-python
numpy
Pillow
gunicorn
